refactor: Product 로직 도메인 안으로 분리
